Refactor ContentView and add dismissal gestures
