+++ Resonans/Components/BottomSheetGallery.swift
@@ -1,0 +1,105 @@
+import SwiftUI
+import Photos
+import UIKit
+
+struct BottomSheetGallery: View {
+    let assets: [PHAsset]
+    let onLastItemAppear: () -> Void
+
+    private let columns: [GridItem] = Array(repeating: .init(.flexible(), spacing: 16, alignment: .center), count: 3)
+
+    var body: some View {
+        let grouped = Dictionary(grouping: assets) { asset in
+            asset.creationDate.map { Calendar.current.startOfDay(for: $0) } ?? Date.distantPast
+        }
+        let sortedDates = grouped.keys.sorted(by: >)
+        LazyVStack(alignment: .leading, spacing: 18) {
+            ForEach(sortedDates, id: \.self) { date in
+                if let items = grouped[date] {
+                    Section(header:
+                        Text(dateFormatted(date))
+                            .font(.system(size: 18, weight: .bold, design: .rounded))
+                            .foregroundStyle(.white.opacity(0.85))
+                            .padding(.leading, 6)
+                            .padding(.bottom, 4)
+                    ) {
+                        LazyVGrid(columns: columns, spacing: 16) {
+                            ForEach(items.indices, id: \.self) { idx in
+                                let asset = items[idx]
+                                let globalIndex = assets.firstIndex(where: { $0.localIdentifier == asset.localIdentifier })
+                                Thumb(asset: asset)
+                                    .onAppear {
+                                        if let gi = globalIndex, gi == assets.count - 1 {
+                                            onLastItemAppear()
+                                        }
+                                    }
+                            }
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    private func dateFormatted(_ date: Date) -> String {
+        let formatter = DateFormatter()
+        formatter.dateStyle = .medium
+        formatter.timeStyle = .none
+        return formatter.string(from: date)
+    }
+
+    private struct Thumb: View {
+        let asset: PHAsset
+        @State private var image: UIImage?
+
+        var body: some View {
+            ZStack {
+                if let image = image {
+                    Image(uiImage: image)
+                        .resizable()
+                        .scaledToFill()
+                        .frame(width: 100, height: 100)
+                        .clipShape(RoundedRectangle(cornerRadius: 26, style: .continuous))
+                        .overlay(alignment: .bottomLeading) {
+                            Text(formatDuration(asset.duration))
+                                .font(.system(size: 15, weight: .bold, design: .rounded))
+                                .padding(8)
+                                .foregroundStyle(.white)
+                                .shadow(color: .black.opacity(0.85), radius: 6, x: 0, y: 2)
+                        }
+                } else {
+                    RoundedRectangle(cornerRadius: 26, style: .continuous)
+                        .fill(Color.white.opacity(0.08))
+                        .frame(width: 100, height: 100)
+                        .overlay(
+                            RoundedRectangle(cornerRadius: 26, style: .continuous)
+                                .strokeBorder(Color.white.opacity(0.15), lineWidth: 1)
+                        )
+                        .onAppear {
+                            if image == nil {
+                                loadThumbnail()
+                            }
+                        }
+                }
+            }
+        }
+
+        private func loadThumbnail() {
+            let manager = PHCachingImageManager()
+            manager.requestImage(for: asset,
+                                 targetSize: CGSize(width: 200, height: 200),
+                                 contentMode: .aspectFill,
+                                 options: nil) { result, _ in
+                image = result
+            }
+        }
+
+        private func formatDuration(_ duration: Double) -> String {
+            let totalSeconds = Int(duration.rounded())
+            let minutes = totalSeconds / 60
+            let seconds = totalSeconds % 60
+            return String(format: "%02d:%02d", minutes, seconds)
+        }
+    }
+}
+
