+++ Resonans/Models/GalleryItem.swift
@@ -1,0 +1,9 @@
+import UIKit
+
+struct GalleryItem: Identifiable {
+    let id: String
+    var thumbnail: UIImage?
+    let duration: String
+    let creationDate: Date
+}
+
