+++ Resonans/Models/RecentItem.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+struct RecentItem: Identifiable {
+    let id = UUID()
+    let title: String
+    let duration: String
+}
+
